showLumi changed to ShowLumi
RegionLabelX     X position of text relative to canvas
RegionLabelY     Y position of text relative to canvas
RegionLabelText  Text to be inserted
LumiX            X position of text relative to canvas if not specified defaults to default position if ShowLumi = KTRUE
LumiY            Y position of text relative to canvas if not specified defaults to default position if ShowLumi = KTRUE
XErrorSize       Defaulting to 0 X error bar size in bin width fraction
IntegerStyle     Defaulting to kFALSE sets to integer style, convenient for njets, needs to be fixed for higher integer values
ArrowX           X position of the arrow in units of the X axis of the channel concerned
ArrowY           Y position of the arrow in units of the Y axis of the channel concerned
ArrowEnd         X position of the end of the arrow in units of the X axis of the channel concerned
ArrowAngle       Defaulting to 40. Arrow Cosmetics
ArrowWidth       Defaulting to 2. Arrow Cosmetics
ArrowColor       Defaulting to 1 (kBlack)

Lower pad if Data/MC shows error bars even if data point outside of canvas